pyproject.toml
[tool.poetry]
name = "svi-smc"
version = "0.1.0"
authors = ["Faisal Jayousi"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.24.0"
numba = "0.59.0"
scipy = "^1.11.0"
pyyaml = "^6.0"
matplotlib = "^3.8.0"
llvmlite = "0.42.0"
pydantic = "^2.12.5"
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "black (>=25.12.0,<26.0.0)",
    "ruff (>=0.14.10,<0.15.0)"
]

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "I"]
src = ["."]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
